Added authorization header

+++ src/main/java/com/autocluck/tweet_scheduler/controller/TweetResource.java
@@ -4,8 +4,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
@@ -21,18 +23,30 @@ public class TweetResource {
 	@Autowired
 	private TweetService tweetService;
 	
+	//GET
+	@GetMapping(path="/health")
+	public String healthCheck() {
+		return "All Good";
+	}
+	
 	//POST
 	@PostMapping(path="/save", consumes = "application/json", produces = "application/json")
-	public ResponseEntity<TweetResponse> saveTweet(@RequestBody CreateTweetRequest request){
+	public ResponseEntity<TweetResponse> saveTweet(@RequestHeader(name = "Authorization", required= true)String authKey, @RequestBody CreateTweetRequest request){
+		if(!tweetService.isAuthorized(authKey)) {
+			return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+		}
 		tweetService.saveTweet(request);
 		return new ResponseEntity<>(HttpStatus.OK);
 	}
 	
-	//PUT
 	
 	//DELETE
 	@DeleteMapping(path="/delete")
-	public void deleteTweet(@RequestParam(value = "tweet") String tweet) {
+	public ResponseEntity<TweetResponse> deleteTweet(@RequestHeader(name = "Authorization", required= true)String authKey, @RequestParam(value = "tweet") String tweet) {
+		if(!tweetService.isAuthorized(authKey)) {
+			return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+		}
 		tweetService.deleteTweet(tweet);
+		return new ResponseEntity<>(HttpStatus.OK);
 	}
 }
